% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addS3Class.R
\name{addS3Class}
\alias{addS3Class}
\title{Add a S3 class to an object.}
\usage{
addS3Class(x, class = NULL)
}
\arguments{
\item{x}{The object to add class.}

\item{class}{The class to add.}
}
\value{
The same object with the new class as first class
}
\description{
Add a S3 class as first class to an object.
}
\details{
addS3Class add a S3 class to an object
}
\examples{
print.CF <- function(x) {cat("print.CF ", x)}
result <- "Je suis donc je pense"
result <- addS3Class(result, class="CF")
class(result)
print(result)
result <- addS3Class(result, class=c("ECF", "OCF"))
class(result)
print(result)
}
\author{
Marc Girondot \email{marc.girondot@gmail.com}
}
