% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc.R
\name{asc}
\alias{asc}
\title{Return the codes (in UTF-8) of a string}
\usage{
asc(x)
}
\arguments{
\item{x}{The string to be analyzed}
}
\value{
A vector with ITF-8 codes of a string
}
\description{
Return the codes (in UTF-8) of a string.
}
\details{
asc returns the codes (in UTF-8) of a string
}
\examples{
asc("abcd")
asc("ABCD")
}
\seealso{
Other Characters: \code{\link{chr}}, \code{\link{d}},
  \code{\link{tnirp}}
}
\author{
Based on this blog: http://datadebrief.blogspot.com/2011/03/ascii-code-table-in-r.html
}
