% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeCoordinate.R
\name{ChangeCoordinate}
\alias{ChangeCoordinate}
\title{Return a value in a changed coordinate}
\usage{
ChangeCoordinate(
  x = stop("At least one value to convert must be provided"),
  initial = stop("Set of two values must be provided as references"),
  transformed = stop("Set of two transformed values must be provided")
)
}
\arguments{
\item{x}{value to convert}

\item{initial}{Set of two values in the original system}

\item{transformed}{Set of the two values in the converted system}
}
\value{
A value in the new system
}
\description{
Return a value in a changed coordinate system.
}
\details{
ChangeCoordinate returns a value in a changed coordinate
}
\examples{
ChangeCoordinate(x=c(10, 20), initial=c(1, 100), transformed=c(0, 1))
}
\author{
Marc Girondot \email{marc.girondot@gmail.com}
}
