% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{chr}
\alias{chr}
\title{Return the characters defined by the codes}
\usage{
chr(n)
}
\arguments{
\item{n}{The code to be used to return a character}
}
\value{
A string with characters defined by the codes
}
\description{
Return a string with characters defined by the codes.
}
\details{
chr returns the characters defined by the codes
}
\examples{
chr(65:75)
chr(unlist(tapply(144:175, 144:175, function(x) {c(208, x)})))
}
\seealso{
Other Characters: 
\code{\link{asc}()},
\code{\link{d}()},
\code{\link{tnirp}()}
}
\author{
Based on this blog: http://datadebrief.blogspot.com/2011/03/ascii-code-table-in-r.html
}
\concept{Characters}
