% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.knitr.R
\name{clean.knitr}
\alias{clean.knitr}
\title{Delete temporary files created during knitr compile}
\usage{
clean.knitr()
}
\value{
Nothing
}
\description{
Delete temporary files created during knitr compile in
working directory.\cr
This function works only in UNIX system (LINUX or MacOSX).\cr
}
\details{
clean.knitr deletes temporary files created during knitr compile
}
\examples{
\dontrun{
clean.knitr()
}
}
\author{
Marc Girondot
}

