% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Run a shiny application for basic functions of comparison}
\usage{
compare()
}
\value{
Nothing
}
\description{
Run a shiny application for basic functions of comparison.
}
\details{
compare runs a shiny application for basic functions of comparison
}
\examples{
\dontrun{
library(HelpersMG)
compare()
}
}
\references{
Girondot, M., Guillon, J.-M., 2018. The w-value: An alternative to t- and X2 tests. Journal of Biostatistics & Biometrics 1, 1-3.
}
\seealso{
Other w-value functions: \code{\link{contingencyTable.compare}},
  \code{\link{series.compare}}
}
\author{
Marc Girondot
}
