% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.tz.R
\name{convert.tz}
\alias{convert.tz}
\title{Convert one Date-Time from one timezone to another}
\usage{
convert.tz(x, tz = Sys.timezone())
}
\arguments{
\item{x}{The date-time in POSIXlt or POSIXct format}

\item{tz}{The timezone}
}
\value{
A POSIXlt or POSIXct date converted
}
\description{
Convert one Date-Time from one timezone to another.\cr
Available timezones can be shown using OlsonNames().
}
\details{
convert.tz Convert one Date-Time from one timezone to another
}
\examples{
d <- as.POSIXlt("2010-01-01 17:34:20", tz="UTC")
convert.tz(d, tz="America/Guatemala")
}
\author{
Marc Girondot
}
\seealso{
Function \code{with_tz()} from \code{lubridate} package does the same. I keep it here only for compatibility with old scripts.
}

