% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{d}
\alias{d}
\title{Write an ASCII Representation of an Object}
\usage{
d(x, file = "", control = c("keepNA", "keepInteger", "showAttributes"),
  collapse = ", \\n  ")
}
\arguments{
\item{x}{A named vector object}

\item{file}{either a character string naming a file or a connection. "" indicates output to the console.}

\item{control}{character vector indicating deparsing options. See .deparseOpts for their description.}

\item{collapse}{Characters used to separate values.}
}
\value{
A string
}
\description{
Writes an ASCII text representation of an R object.\cr
It can be used as a replacement of dput() for named vectors.\cr
The controls "keepNA", "keepInteger" and "showAttributes" are utilized for named vectors.
}
\details{
d Write an ASCII Representation of an Object
}
\examples{
d(c(A=10, B=20))
dput(c(A=10, B=20))
}
\seealso{
Other Characters: \code{\link{asc}}, \code{\link{chr}},
  \code{\link{tnirp}}
}
\author{
Marc Girondot
}
