% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growlnotify.R
\name{growlnotify}
\alias{growlnotify}
\title{Send growl notification for MacOs X system.}
\usage{
growlnotify(textinfo = "")
}
\arguments{
\item{textinfo}{Text to display in the growlnotify window}
}
\value{
None
}
\description{
This function was used to send a notification to MacOS user.\cr
Growlnotify being no longer supported in MacOSX, this function will be removed in future releases.
}
\details{
growlnotify sent growl notification for MacOs X systems. No more supported.
}
\examples{
\dontrun{
# If growlnotify is used on a non-mac system, it just quits.
growlnotify("It works if you are on a Mac with GrowlNotify installed!")
}
}
\author{
Marc Girondot
}
