% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCutter.R
\name{iCutter}
\alias{iCutter}
\title{Run a shiny application to fit bone section}
\usage{
iCutter()
}
\value{
Nothing
}
\description{
Run a shiny application to fit bone section
}
\details{
BP runs a shiny application to fit bone section
}
\examples{
\dontrun{
# Not run:
library(HelpersMG)
iCutter()
}
}
\author{
Marc Girondot \email{marc.girondot@gmail.com}
}
