% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invlogit.R
\name{invlogit}
\alias{invlogit}
\title{Return the inverse logit}
\usage{
invlogit(n)
}
\arguments{
\item{n}{The value to inverse to get the probability}
}
\value{
A value
}
\description{
Return the inverse logit.
}
\details{
invlogit returns the inverse logit
}
\examples{
n <- logit(0.5)
invlogit(n)
}
\seealso{
Other logit: 
\code{\link{flexit}()},
\code{\link{logit}()}
}
\author{
Marc Girondot
}
\concept{logit}
