% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.packages.R
\name{list.packages}
\alias{list.packages}
\title{List the installed packages with their locations}
\usage{
list.packages()
}
\value{
A list with the installed packages and their version.
}
\description{
List the installed packages with their locations and version.
}
\details{
list.packages lists the installed packages with their locations
}
\examples{
\dontrun{
library(HelpersMG)
list.packages()
}
}
\author{
Marc Girondot
}
