% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Return the logit}
\usage{
logit(p)
}
\arguments{
\item{p}{The probability}
}
\value{
A value
}
\description{
Return the logit.
}
\details{
logit returns the logit
}
\examples{
n <- logit(0.5)
invlogit(n)
}
\seealso{
Other logit: \code{\link{invlogit}}
}
\author{
Marc Girondot
}
