% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.compareAIC.R
\name{logLik.compareAIC}
\alias{logLik.compareAIC}
\title{Return Log Likelihood generated by FormatCompareAIC}
\usage{
\method{logLik}{compareAIC}(object, ...)
}
\arguments{
\item{object}{A result generated by FormatCompareAIC}

\item{...}{Not used}
}
\value{
The Log Likelihood value for the fitted model with data
}
\description{
Return Log Likelihood generated by FormatCompareAIC
}
\details{
logLik.compareAIC Return Log Likelihood of a fit
}
\examples{
\dontrun{
ED <- FormatCompareAIC(logLik=-140, nobs=100, df=3)
logLik(ED)
}
}
\author{
Marc Girondot
}
\concept{AIC functions}
