% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moon_phase.R
\name{moon_phase}
\alias{moon_phase}
\title{Moon phase based on a date}
\usage{
moon_phase(date = NULL, phase = FALSE)
}
\arguments{
\item{date}{A date in known format}

\item{phase}{If TRUE, a vector of character with NM, FQ, FL LQ will be returned}
}
\value{
Return a value describing the moon phase:\cr
0 and 100 are full moon, 50 is new moon, 25 last quarter and 75 first quater
}
\description{
The script gives an index (base 100) that represents moon phase. If the value lays between:\cr
0 and 1.6931595 or 98.3068405 and 100, it is full moon,\cr
23.3068405 and 26.6931595, last quarter,\cr
48.3068405 and 51.6931595, new moon,\cr
73.3068405 and 76.6931595, first quarter\cr
When phase is set to TRUE, a character representing the moon phase is returned.
}
\details{
moon_phase calculates the moon phase based on a date.
}
\examples{
library("HelpersMG")
moon_phase(as.Date("2001-12-31"))
moon_phase(as.Date("14/04/2010", "\%d/\%m/\%Y"))
moon_phase(as.Date("22/06/07", "\%d/\%m/\%y"))
moon_phase(seq(from=as.Date("2012-03-01"),
		to=as.Date("2012-04-15"), by="days"))
moon_phase(seq(from=as.Date("2012-03-01"),
		to=as.Date("2012-04-15"), by="days"), phase=TRUE)
}
\author{
Marc Girondot \email{marc.girondot@u-psud.fr}
}
\keyword{Lunar}
\keyword{Lune}
\keyword{Moon}

