% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwd.R
\name{openwd}
\alias{openwd}
\title{Open a finder window with current working directory in MacOS X}
\usage{
openwd(directory = getwd())
}
\arguments{
\item{directory}{The directory you want to open}
}
\value{
A vector with the list of files.
}
\description{
This function opens a finder window with directory files 
in MacOS X. It has not been tested in Windows. In linux, it just returns the 
list of files in directory.\cr
By defaut, it uses the current working directory.
}
\details{
openwd will open a finder window with current working directory
}
\examples{
\dontrun{
openwd()
}
}
\author{
Marc Girondot
}
