% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_add.R
\name{plot_add}
\alias{plot_add}
\title{Add a plot to a previous one}
\usage{
plot_add(...)
}
\arguments{
\item{...}{Parameters for plot()}
}
\value{
Nothing
}
\description{
To plot data, just add use it as a normal plot. It will plot 
the new data without axes, or labels for axes.\cr
This function is complementary to matlines() and matpoints() from package graphics.
}
\details{
plot_add adds a plot to a previous one
}
\examples{
\dontrun{
plot(x=1:100, y=sin(1:100), type="l", bty="n", xlim=c(1,200), xlab="x", ylab="y")
plot_add(x=1:200, y=cos(1:200), type="l", bty="n", col="red")
}
}
\author{
Marc Girondot
}
\seealso{
Other plot and barplot functions: \code{\link{ScalePreviousPlot}},
  \code{\link{barplot_errbar}}, \code{\link{plot_errbar}}
}

