% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSnbinom.R
\name{pSnbinom}
\alias{pSnbinom}
\title{Distribution function for the sum of random variable with negative binomial distributions.}
\usage{
pSnbinom(q = stop("At least one quantile must be provided"), size = NULL,
  prob = NULL, mu = NULL, lower.tail = TRUE, log.p = FALSE,
  infinite = 100)
}
\arguments{
\item{q}{vector of quantiles.}

\item{size}{target for number of successful trials, or dispersion parameter (the shape parameter of the gamma mixing distribution). Must be strictly positive, need not be integer.}

\item{prob}{probability of success in each trial. 0 < prob <= 1.}

\item{mu}{alternative parametrization via mean.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{infinite}{Number of maximal iterations; check different values to determine the error in estimation.}
}
\value{
pSnbinom returns distribution function
}
\description{
Distribution function for the sum of random variable with negative binomial distributions.
}
\details{
pSnbinom returns the distribution function for the sum of random variable with negative binomial distributions
}
\examples{
alpha <- c(2.1, 2.05, 2)
mu <- c(10, 30, 20)
p <- pSnbinom(q=10, size=alpha, mu=mu, lower.tail = TRUE)
}
\author{
Marc Girondot
}
\seealso{
Other Distribution of sum of random variable with negative binomial distributions: \code{\link{dSnbinom}},
  \code{\link{qSnbinom}}, \code{\link{rSnbinom}}
}

