% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_decimals.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Return a number as character with specified number of decimals}
\usage{
specify_decimal(x, decimals = 3, decimal.point = ".")
}
\arguments{
\item{x}{The numbers to be formated}

\item{decimals}{Number of decimals to print}

\item{decimal.point}{Character to be used as decimal point}
}
\value{
A character
}
\description{
Return a number as character with specified number of decimals.
}
\details{
specify_decimals format a number with specified number of decimals
}
\examples{
specify_decimal(x=pi, decimals=3, decimal.point=".")
specify_decimal(x=c(pi, exp(1)), decimals=3, decimal.point=",")
}
\author{
Marc Girondot
}
