% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{wget}
\alias{wget}
\title{Download a file from internet and save it locally}
\usage{
wget(url = stop("At least one internet adress is required"), ...)
}
\arguments{
\item{url}{The url where to download file}

\item{...}{The parameters send to download.file()}
}
\value{
Nothing
}
\description{
Download a file from internet and save it locally. This function is a wrapper for
download.files() that keep the name identical and can get several files at once.
}
\details{
wget download a file from internet and save it locally
}
\examples{
\dontrun{
library(HelpersMG)
# Save locally the files send in the parameter url
wget(c("https://cran.r-project.org/web/packages/HelpersMG/HelpersMG.pdf", 
         "https://cran.r-project.org/web/packages/embryogrowth/embryogrowth.pdf"))
}
}
\author{
Marc Girondot
}
