% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Run a shiny application for basic functions of comparison}
\usage{
compare()
}
\value{
Nothing
}
\description{
Run a shiny application for basic functions of comparison
}
\details{
compare runs a shiny application for basic functions of comparison
}
\examples{
\dontrun{
library(HelpersMG)
compare()
}
}
\author{
Marc Girondot
}
