% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growlnotify.R
\name{growlnotify}
\alias{growlnotify}
\title{Send growl notification for MacOs X system.}
\usage{
growlnotify(textinfo = "", help = FALSE)
}
\arguments{
\item{textinfo}{Text to display in the growlnotify window}

\item{help}{If TRUE, an help is displayed}
}
\value{
None
}
\description{
This function is used to send a notification to MacOS user.
}
\details{
growlnotify send growl notification for MacOs X systems.
}
\examples{
# If growlnotify is used on a non-mac system, it just quits.
growlnotify("It works if you are on a Mac with GrowlNotify installed!")
}
\author{
Marc Girondot
}

